+++ src/main/kotlin/interfaces/IEntry.kt
@@ -7,7 +7,7 @@ interface IEntry {
   /**
    * The entry's unique identifier.
    */
-  var uID: Int
+  var uID: Long
 
   /**
    * Used to initialize an entry, e.g. setting the unique identifier when creating a new entry.
