+++ src/main/kotlin/modules/m2/logic/ContactController.kt
@@ -36,8 +36,8 @@ class ContactController : IModule {
     val mutex = Mutex()
   }
 
-  private suspend fun saveEntry(knowledge: Knowledge): Int {
-    var uID: Int
+  private suspend fun saveEntry(knowledge: Knowledge): Long {
+    var uID: Long
     mutex.withLock {
       uID = save(knowledge)
     }
