+++ src/main/kotlin/modules/m4storage/misc/ItemStorages.kt
@@ -10,5 +10,5 @@ import modules.m4storage.logic.ItemStorageManager
 @ExperimentalSerializationApi
 @Serializable
 data class ItemStorages(val type: ItemStorageManager.CategoryType) {
-  val storages = mutableMapOf<Int, ItemStorage>()
+  val storages = mutableMapOf<Long, ItemStorage>()
 }
