+++ src/main/kotlin/modules/mx/LastChange.kt
@@ -3,4 +3,4 @@ package modules.mx
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class LastChange(var uID: Int, var unixHex: String, var user: String)
+data class LastChange(var uID: Long, var unixHex: String, var user: String)
